Port src/marks/tree from JS to TS (stage 1)
Inline hand-written .d.ts types into src/marks/tree.js to produce a
single src/marks/tree.ts. No behaviour changes; runtime is identical.
